rename twitter_bot to twitter_bot_func_app
