+++ get_users/get_users.py
@@ -5,13 +5,13 @@ import json
 import logging
 import azure.functions as func
 
-from twitter_bot.brokers import TwitterBroker
-from twitter_bot.data import PoliticianDataService, WikipediaDataBroker, Politician
-from twitter_bot.model import TwitterUser
-from twitter_bot.enums import PoliticianType
-from twitter_bot.serialization import Encoder
-from twitter_bot.utils.functions import load_json
-from twitter_bot.utils.constants import REPRESENTATIVES_WIKI_URL, SENATORS_WIKI_URL, \
+from twitter_bot_func_app.brokers import TwitterBroker
+from twitter_bot_func_app.data import PoliticianDataService, WikipediaDataBroker, Politician
+from twitter_bot_func_app.model import TwitterUser
+from twitter_bot_func_app.enums import PoliticianType
+from twitter_bot_func_app.serialization import Encoder
+from twitter_bot_func_app.utils.functions import load_json
+from twitter_bot_func_app.utils.constants import REPRESENTATIVES_WIKI_URL, SENATORS_WIKI_URL, \
     TOTAL_NUM_REPRESENTATIVES, TOTAL_NUM_SENATORS
