+++ twitter_bot_func_app/brokers/azure_cloud_broker.py
@@ -8,7 +8,7 @@ import requests
 from azure.identity import DefaultAzureCredential
 from azure.core.credentials import AccessToken
 
-from twitter_bot.utils.constants import AZURE_CONFIG_FILEPATH, SECRETS_FILEPATH, LOCAL_EXECUTION
+from twitter_bot_func_app.utils.constants import AZURE_CONFIG_FILEPATH, SECRETS_FILEPATH, LOCAL_EXECUTION
 
 class AzureCloudBroker():
     """
