+++ twitter_bot_func_app/brokers/azure_text_analytics_broker.py
