+++ twitter_bot_func_app/enums/politician_type.py
