+++ twitter_bot_func_app/enums/sentiment.py
