+++ twitter_bot_func_app/utils/functions.py
